+++ src/Cpnucleo.RazorPages/Pages/Impedimento/Remover.cshtml.cs
@@ -24,10 +24,17 @@ namespace Cpnucleo.RazorPages.Pages.Impedimento
         {
             try
             {
-                Impedimento = await _impedimentoService.ConsultarAsync(Token, id);
+                var result = await _impedimentoService.ConsultarAsync(Token, id);
+
+                if (!result.sucess)
+                {
+                    ModelState.AddModelError(string.Empty, $"{result.code} - {result.message}");
+                    return Page();
+                }
+
+                Impedimento = result.response;
 
                 return Page();
-
             }
             catch (Exception ex)
             {
@@ -40,7 +47,28 @@ namespace Cpnucleo.RazorPages.Pages.Impedimento
         {
             try
             {
-                await _impedimentoService.RemoverAsync(Token, Impedimento.Id);
+                if (!ModelState.IsValid)
+                {
+                    var result = await _impedimentoService.ConsultarAsync(Token, Impedimento.Id);
+
+                    if (!result.sucess)
+                    {
+                        ModelState.AddModelError(string.Empty, $"{result.code} - {result.message}");
+                        return Page();
+                    }
+
+                    Impedimento = result.response;
+                    
+                    return Page();
+                }
+
+                var result2 = await _impedimentoService.RemoverAsync(Token, Impedimento.Id);
+
+                if (!result2.sucess)
+                {
+                    ModelState.AddModelError(string.Empty, $"{result2.code} - {result2.message}");
+                    return Page();
+                }
 
                 return RedirectToPage("Listar");
             }
