+++ src/Cpnucleo.RazorPages/Pages/Projeto/Remover.cshtml.cs
@@ -24,7 +24,15 @@ namespace Cpnucleo.RazorPages.Pages.Projeto
         {
             try
             {
-                Projeto = await _projetoService.ConsultarAsync(Token, id);
+                var result = await _projetoService.ConsultarAsync(Token, id);
+
+                if (!result.sucess)
+                {
+                    ModelState.AddModelError(string.Empty, $"{result.code} - {result.message}");
+                    return Page();
+                }
+
+                Projeto = result.response;                
 
                 return Page();
             }
@@ -39,7 +47,28 @@ namespace Cpnucleo.RazorPages.Pages.Projeto
         {
             try
             {
-                await _projetoService.RemoverAsync(Token, Projeto.Id);
+                if (!ModelState.IsValid)
+                {
+                    var result = await _projetoService.ConsultarAsync(Token, Projeto.Id);
+
+                    if (!result.sucess)
+                    {
+                        ModelState.AddModelError(string.Empty, $"{result.code} - {result.message}");
+                        return Page();
+                    }
+
+                    Projeto = result.response;
+                    
+                    return Page();                    
+                }
+
+                var result2 = await _projetoService.RemoverAsync(Token, Projeto.Id);
+
+                if (!result2.sucess)
+                {
+                    ModelState.AddModelError(string.Empty, $"{result2.code} - {result2.message}");
+                    return Page();
+                }
 
                 return RedirectToPage("Listar");
             }
